Removing test run dependency on the dashboard database
The tests can still run, and by default will try to connect to the dashboard database and push the results. However, if the connection fails, the results will simply be printed to screen and the connection failure exception will be handled by an optional user supplied action.

This allows us to maintain compatibility with the current behavior while not getting blocked to run these tests when the dashboard database isn't present. This is also a step forward in running these tests from platforms other than Windows.