Benchmark database connectivity check
This enables a very simple level of abstraction where the benchmark result processor takes care of verifying if a connection can be made against a database before proceeding to save the results. This check is only performed once per execution. If a connection can be established, but an error occurs during data insertion, the exception will still pop-up to the caller the same way it has always done.

The change enables the benchmark framework to run without requiring a localdb instance to be available (such as in Linux systems). Tests that don't explicitly ask for a connection to localdb should be able to run everywhere now. This resolves issue #29.